Use new format in blocks-cli
Summary:
D7129 introduced the [new backend blocks response payload](https://airtable.quip.com/McMGAlUCDHiS/RFC-Backend-blocks-response-payload) in Hyperbase. This is a similar change but in blocks-cli.

This diff is part 2 in a four-step plan:

1. Update Hyperbase to support old and new response formats ✅
2. Update blocks-cli to support new response format
3. Migrate all existing blocks to the new response format (perhaps by building a helper library, or perhaps another kind of helper)
4. Remove the parsing of old response formats from Hyperbase

Significantly, it has no tolerance for the old blocks format. If you're developing a backend block with blocks-cli, it won't work with the old format. This might be a little abrasive, but (1) I'll make an announcement after this lands with help (2) the changes aren't radically different (3) it should hasten our progress towards the new world.

Test Plan:
No QA is needed for this change.

Tested this by adding tests and making sure this response worked with a development block...

```
lang=javascript
{
  statusCode: 200,
  headers: ['X-Foo', 'boo'],
  body: Buffer.from('hello world').toString('base64'),
  bodyFormat: 'base64',
}
```

...and that this one did not:

```
lang=javascript
{
  statusCode: 200,
  headers: {'X-Foo': 'boo'},
  body: 'hello world',
}
```

Reviewers: kasrak, jbbakst

Reviewed By: kasrak, jbbakst

Differential Revision: https://airtable.phacility.com/D7181