Refactor: change how commands are registered/run
Summary:
Creates an abstraction over the concept of an action so we're not bloating `run_block_cli` every time we add an action. This will make it easier to add new actions in the future.

I also moved them into their own directory, since the `lib` directory is a nightmare to navigate right now

Test Plan: Test that all commands work and that help output looks right

Reviewers: kasrak, EvanHahn, richsinn

Reviewed By: kasrak, richsinn

Differential Revision: https://airtable.phacility.com/D8678