Move run into own module + higher-order function for running command
Summary: Moves `block run` into its own command module. Also adds a higher-order function to run a command with a given name (assumes that there is a file named after the command that exports an obj with an `runCommandAsync` function).

Test Plan: Verify that run, clone, and push all work

Reviewers: kasrak, richsinn

Reviewed By: richsinn

Differential Revision: https://airtable.phacility.com/D8696