Generate frontend bundles
Summary:
Adds a new command to "build" your block. Right now, it generates the frontend bundle but not the backend deployment package (wanted to minimize the size of the initial diff, since it was already pretty big).

Currently, this attempts to replicate behavior from Hyperbase's block builder. In the future, we can start adjusting how local builds work to offer more flexibility to block developers.

Test Plan: Run `block build` and verify that a js bundle is generated in the `build_artifacts` directory (in whatever `tmp` dir the build is generated in)

Reviewers: kasrak, somehats

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D8893