transpile src to dist with babel
Summary:
- Transpiles src to dist with babel
- Copies source files to dist with flowtypes in tact for shipping flow version
- Documents npm commands

babel.config.js is set up slightly differently to other blocks projects - it uses the `useBuiltIns: 'usage'` and `@babel/plugin-transform-runtime` to make polyfills etc. available. When used together, these options:

- only include those polyfills that are needed by the transpile target, and only when it sees them used in the source
- scopes them locally, rather than polluting the global namespace as `require('core-js')` and similar usually do

Together, this means that we can use all the modern JS we want to in the SDK and be confident that a minimal set of polyfills will be included in every project that pulls in the SDK.

Depends on D9224

Test Plan:
- Run `yarn run build`. It should build the project in `dist`
- Run `yarn run watch`. It should build the project in `dist`, and any source files that get changed should stay up to date.

Reviewers: kasrak, EvanHahn

Reviewed By: kasrak, EvanHahn

Differential Revision: https://airtable.phacility.com/D9227