remove airtable_interface and liveapp_interface in favor of injected airtableInterface
Summary:
See D9276 for the hyperbase counterpart to this diff.

Currently, the injected airtable interface is identical to the removed `airtable_interface_frontend` except for an additional `initialData` field used for `sdkInitData`.

Test Plan: With this checked out and both D9276 & D9277 running in hyperbase, try running the chart block locally on branch `chart-newSdk`.

Reviewers: kasrak, jbbakst, EvanHahn

Reviewed By: kasrak

Differential Revision: https://airtable.phacility.com/D9278