move 'src/arc' directory to root 'arc'
Summary:
Depends on D9419

Why
===
In preparation for the new `blocks-cli` project structure, move out the phab related configuration from the `src` directory to the root.
Reason: The `arc` folder is a "configuration/tooling" folder for our own internal development process. We should move it out of `src` and maintain it at the root.

TODO: The `hyperbase` repo also has a `.arcunit` config file, but it's missing in `blocks-cli`. Therefore, the configuration in the `arc` directory is actually **not** running at the moment. Some of the checks that is handled by this existing configuration (which is actually not being run in this repo) is to make sure that the developer's local branch has been pushed to origin (in github) - this is so that Airbuild can run. The `blocks-cli` repo actually doesn't require such workflow right now, because there is no Airbuild or CI integration with `blocks-cli`. Therfore, I'm holding off on creating an `.arcunit` file since our processes will differ slightly from the `hyperbase` repo development processes right now, but deciding to keep these configurations here unless we're absolutely clear we want to remove them.

See D9279 for spike exploration for the new `blocks-cli` project structure.

What
====
Move `src/arc` directory to root `arc`.

Test Plan: No QA Needed

Reviewers: jbbakst, somehats, kasrak

Reviewed By: somehats

Subscribers: #blocks

Differential Revision: https://airtable.phacility.com/D9441