flow-ify APIClient part 2 - make blockId optional
Summary:
Depends on D9480

`APIClient` was flow-ified in D9480, but it missed making `opts.blockId` optional in the constructor interface.

This check was missed because the other files instantiating `APIClient` are currently not flow-ified.

The current constructor interface for `APIClient` is not great because both `opts.blockId` and `opts.blockInstallationId` are optional, but at least one must to be present (depends on which API it plans to call).

Short-term is to fix the flow-typing to fit the current model. Broader refactor will come later.

Test Plan: No QA Needed

Reviewers: EvanHahn, jbbakst, kasrak

Reviewed By: EvanHahn

Subscribers: #blocks

Differential Revision: https://airtable.phacility.com/D9495