blocks-cli: Set up mock airtable.com tests by testing APIClient.prototype.fetchBlockAsync
Summary:
This tests `APIClient.prototype.fetchBlockAsync` and sets up everything necessary to do that.

`7132` is used as a local port throughout, which is just a random port (I got it from `echo "$RANDOM"`). In the future, we may wish to make this more like "get me an open port", which would require a little more plumbing because that port number would be dynamic.

Note that `no-loop-func` has been removed from Hyperbase, so we remove it here, too.

Test Plan: QA can skip

Reviewers: richsinn, jbbakst

Reviewed By: richsinn

Differential Revision: https://airtable.phacility.com/D9508