Add type for `required` attribute of `UI.Input`
Summary:
Very similar to D9647.

Some blocks were sending required as a string, which was triggering a warning from React. This adds Flow and prop types to ensure that doesn't happen in the future.

We should probably apply this change to the new SDK, too—not sure if that will get picked up automatically.

Test Plan: QA can skip

Reviewers: somehats, s2tephen, jay-ransijn

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D9669