When publishing blocks_cli to npm, omit unnecessary files (like test/ and flow-typed/)
Summary:
This uses the ["files" key of package.json](https://docs.npmjs.com/files/package.json#files) to include only files we need.

It may seem strange that we're including `/src/` instead of `/transpiled/`. This is deliberate, because we compile the source //on install//, not //on publish//. This is something we should fix, but is a separate task.

Tested this by:

1. Running `npm pack`
2. Copying the resulting tarball (`airtable-blocks-cli-0.0.1.tgz`) to a temporary directory
3. `npm install ./airtable-blocks-cli-0.0.1.tgz` from that temporary directory
4. Ran `"$(npm bin)/block" --help` and verified that the help screen appeared

Test Plan: QA can skip

Reviewers: somehats, kasrak, richsinn, jbbakst

Reviewed By: richsinn

Subscribers: #blocks

Differential Revision: https://airtable.phacility.com/D9746