Add license to blocks-cli
Summary:
We'll need a license before we publish this to npm (even privately).

The most valuable comments on this diff aren't "did you set up the files correctly" (although that's helpful), but are "is the MIT license the right one"? It's the same one we use for Airtable.js, but I could see it being the wrong one for our use case.

Test Plan: QA can skip

Reviewers: kasrak, jbbakst

Reviewed By: kasrak

Differential Revision: https://airtable.phacility.com/D9752