cli: Non-tls HTTP '/ping' route
Summary:
Depends on D9773

Why
====
We want to create a non-TLS HTTP `/ping` route to use as a heuristic to determine if the the block frame's browser has enabled running localhost in "unsafe" mode.

What
=====
* Create an `http` server (to be used on the same express app as the `https` server)
* Add checks for `req.secure` in route registration.
* Allow `/ping` route to pass through in both secure and non-secure mode.
Tested locally
----------------
* run `block run --local`
* open block frame in local dev mode
* open browser console
* run the following commands in the browser console:
** `await fetch('https://localhost:8000/__runFrame/ping', {method: 'HEAD'});` => should PASS
** `await fetch('http://localhost:80/__runFrame/ping', {method: 'HEAD'});` => should PASS
** `await fetch('https://localhost:8000/__runFrame/bundle.js', {method: 'GET'});` => should PASS
** `await fetch('http://localhost:80/__runFrame/bundle.js', {method: 'GET'});` => should 404

TODO: There will be a follow-up diff in hyperbase to use this non-TLS `/ping` as a heuristic to determine if the browser has enabled localhost in "unsafe" mode.

Test Plan: No QA Needed

Reviewers: kasrak, jbbakst, somehats

Reviewed By: kasrak

Subscribers: #blocks

Differential Revision: https://airtable.phacility.com/D9826