avoid generating different tarballs for the same version
Summary:
Weird issue where as we were generating tarballs of the same name and version, yarn's cache was getting into weird states.

This (giant hack) assigns a unique version to each tarball. It's enough of a weird workaround to suggest we should probably rethink how we're doing this altogether, but it's enough to get things working for now.

Test Plan: - block run with --sdk-repo

Reviewers: jbbakst, richsinn

Reviewed By: jbbakst, richsinn

Subscribers: #blocks

Differential Revision: https://airtable.phacility.com/D9957