remove 'block_backend_wrapper' dir for Alpha
Summary:
We removed backend support for the "alpha" version of `blocks-cli`. Most of the backend related logic was already removed except the `block_backend_wrapper` folder; removing it in this diff.

We will add this back in when it comes time to implement the new backend.

Tested locally for regressions.

Test Plan: No QA Needed

Reviewers: jbbakst

Reviewed By: jbbakst

Differential Revision: https://airtable.phacility.com/D9963