Add ESLint rules for JSDoc
Summary:
This installs and configures [eslint-plugin-jsdoc](https://github.com/gajus/eslint-plugin-jsdoc). This will help us as we add automatic API documentation.

I opted to make things errors instead of warnings, but used warnings when the change would be too distruptive. (For example, I made the `require-jsdoc` a warning, not an error.)

The most conversial part of this diff involves top-level `await`s. I added `allowAwaitOutsideFunction` to our `babel.config`, and in order to pick it up, we had to upgrade `babel-eslint` to `11.0.0-beta.0`. I'm comfortable using a beta for something like Babel when it's only used for linting, but I could see that being too bleeding-edge for us.

Test Plan: QA can skip

Reviewers: somehats, jbbakst, kasrak

Reviewed By: somehats, jbbakst

Differential Revision: https://airtable.phacility.com/D9942