Hacky: Syntax Error in frame + autoreload on syntax fix
Summary:
Depends on D9987

Why
===
This diff contains a hacky solution for:
1. Displaying syntax errors in the block frame itself.
2. Auto-loading after the syntax error has been fixed.

This solution is hacky for several reasons, but it localizes changes to the `blocks-cli` tool itself (doesn't require changes to hyperbase). A long term solution would bubble up the syntax error to hyperbase, and hyperbase's block frame would handle showing in the "Error Overlay"; Hyperbase would also eventually be responsible for auto-reloading somehow using the `poll` endpoint.

What
====
* Upon syntax error, converts the "ansi-encoded, syntax-error stack trace" to HTML.
** Introduced external dependency `ansi-to-html` to handle the ansi to HTML conversion.
** Return this HTML, along with the `BUNDLE_ERROR` error code in the response body of the `poll` request.
* Modify `generate_block_client_wrapper` to show a hacky overlay error div if HTTP response is 500, and an error code of `BUNDLE_ERROR` is detected.
** Make another recursive request to `poll` to keep the polling connection alive.

This solution does NOT handle trying to start a local `block run` if there already was a syntax error to begin with. It only handles if a syntax error occurs once the block is already running.

Tested locally:
* `block run`
* In Hyperbase, click "Edit block" -> input in the bundle URL (i.e `https://localhost:8000`)
* In the block's user code, make a syntax error. Save the file
* The frame should show an overlay with the syntax stack trace
** The `poll` request should've received an HTTP 500 status with a response body containing both the error code and the syntax error stack trace.
** screenshot: {F121082}
* Fix the syntax error in the block's user code file and click save. The block should auto-reload successfully.

Test Plan: No QA Needed

Reviewers: jbbakst, kasrak

Reviewed By: jbbakst

Differential Revision: https://airtable.phacility.com/D9992