Fix block directory path resolution on Windows
Summary:
`/` was hard-coded as the root of the filesystem, but that's not the case on Windows, where it's something like `C:\`.

I also wrote tests for this. We were caching the blocks directory path (originally introduced in 7b9f25d8fb3890e3d6c49b9ba4f6de048e8d4316) which made it near impossible to test more than one invocation of the function, so I removed the caching. This has two other benefits:

1. If `process.cwd()` ever changes (perhaps because we `process.chdir`), the old block directory path will be updated
2. If files are removed while `block` is running (perhaps `block.json` is deleted), the tool will error instead of having strange behavior

However, it's marginally slower because we're hitting the fileystem every time. I think this is negligible, but I may be missing why we cached the result in the first place.

I also stopped `process.exit`ing when no block directory is found, and instead `throw`. It looks like we used to throw, but that was also changed in 7b9f25d8fb3890e3d6c49b9ba4f6de048e8d4316—again, I'm not sure of the original motivation there. This change prevents an ESLint error, and it seems to behave the same when I tested it in the command line.

Fixes https://airtable.com/tblirJUiB3DryM26j/recA3sGEWAcCCQWaR

Test Plan: QA can skip

Reviewers: kasrak, jbbakst, richsinn

Reviewed By: richsinn

Differential Revision: https://airtable.phacility.com/D10076