fix SVGElement flow error
Summary:
Generally, we need to keep `.flowconfig` as vanilla as possible - flow won't use our config when checking the SDK when it's installed in another project, so we can't rely on `.flowconfig` to prevent errors.

Instead of using `.flowconfig` to declare a global SVGElement (which only works in this repo - it errors when the SDK is installed in a project), we import an interface that matches the same type.

Test Plan: - no flow errors in this repo or projects using the SDK

Reviewers: s2tephen

Reviewed By: s2tephen

Differential Revision: https://airtable.phacility.com/D10124