Fix ESLint error that prevents commits
Summary:
On `master`, this fails:

```
lang=sh
yarn format && yarn lint:quiet
```

This is causing commits to blow up.

This is because the `// eslint-disable-line` comments are being moved by Prettier, which causes problems. I got rid of the `eslint-disable-line` comments and used a `/* eslint-disable */` block.

Alternatives that we didn't do:

1. `// eslint-disable-next-line` doesn't work because then the JSDoc comments get "unlinked". Doing that causes the `jsdoc/require-jsdoc` to give a warning.
1. `// prettier-ignore` doesn't work for the same reason.
1. `// prettier-ignore-start` also works, but that would prevent a large chunk of code from being formatted.

Test Plan: QA can skip

Reviewers: kasrak, jbbakst, somehats

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D10129