Add changelog
Summary:
This is a public-facing changelog that aims to keep both internal and external consumers of the SDK aware of ongoing changes to the project. It can also help contributors understand where we're at.

The changelog should follow the format set out in https://keepachangelog.com/en/1.0.0/. Details of when and how to update the changelog are in the file.

When the project has its version bumped, [[ https://github.com/SomeHats/changelog-publish | changelog-publish ]] will automatically update the changelog. That repo should be reviewed as well as part of this.
It will:
- Rename the 'Unreleased' section to the new version and add the date
- Link the new version to a diff between it and the previous version on GitHub
- Add a new empty 'Unreleased' section at the top and link it to a diff between the version you just released and `HEAD`

Once this is approved I'll open up a similar diff in blocks-cli

Test Plan: -

Reviewers: EvanHahn, kasrak, jbbakst

Reviewed By: EvanHahn, jbbakst

Differential Revision: https://airtable.phacility.com/D10147