add changelog
Summary:
This is a public-facing changelog that aims to keep both internal and external consumers of the CLI aware of ongoing changes to the project. It can also help contributors understand where we're at.

The changelog should follow the format set out in https://keepachangelog.com/en/1.0.0/. Details of when and how to update the changelog are in the file.

When the project has its version bumped, [[ https://github.com/SomeHats/changelog-publish | changelog-publish ]] will automatically update the changelog.
It will:
- Rename the 'Unreleased' section to the new version and add the date
- Link the new version to a diff between it and the previous version on GitHub
- Add a new empty 'Unreleased' section at the top and link it to a diff between the version you just released and `HEAD`

Right now, this is set up to generate links pointing to github.com/airtable/blocks-cli. That repo doesn't exist yet, but is where the cli source will soon be published.

See also D10147

Test Plan: -

Reviewers: richsinn, jbbakst

Reviewed By: richsinn, jbbakst

Differential Revision: https://airtable.phacility.com/D10243