cli: implement 'bundle.js' endpoint to return syntax err
Summary:
Why
===
Handle bundle error on "first load" (or reload). Currently, if there is a bundle error, then on the block frame's "first load" it will show an unhelpful "missing `_airtableReact`" error. This is because the bundle file is empty and the `_airtableReact` variable is missing in the global scope.

This diff refactors the `/bundle.js` endpoint to return a `422` HTTP status code and the bundle error stack (ansi to HTML form) in the response body, if there was a bundle failure. In a subsequent diff, changes will be made to hyperbase to support consuming this error stack.

What
====
* In `block_server.js`:
** In the `/bundle.js` middleware, catch any errors, and return a `500` HTTP status code and the related error stack if it's related to a `BUNDLE_ERROR`.
* Refactor the bundle status instance var from `_bundlePromiseIfExists: Promise<void> | null;` to `_bundleStateData: BundleStateData;`- this helps us bubble up the bundle error information to the request handler for the `/bundle.js` endpoint.

Test Plan: Tested manually

Reviewers: jbbakst, kasrak, somehats, minicat

Reviewed By: jbbakst, somehats

Differential Revision: https://airtable.phacility.com/D10265