add useGlobalConfig hook
Summary:
Adds a useGlobalConfig hook, following the same pattern as useViewport - we watch everything and return the instance.

This is an alternative to the `useGlobalConfigPath` approach we explored in D9937 - we offer a single hook that gives us access to the whole of globalConfig rather than selecting a specific path. This means we don't have to provide some alternative API for permission handling.

Test Plan: Created test block

Reviewers: jbbakst, kasrak

Reviewed By: kasrak

Differential Revision: https://airtable.phacility.com/D10247