Helpers to read/write API key from/to config files
Summary:
As a first step towards the new API key storage system documented in https://airtable.quip.com/abvAAZRmrsdM/June-2019-API-Key-Storage-in-blocks-cli, implement helpers that read and write API key to the new storage locations. As per the doc, we decided to keep it as simple as possible for this "v0", only supporting two locations and a single attribute that can be read/written to the config.

Will plug it in in a following diff.

Heavily borrows from https://airtable.phacility.com/D9616?vs=on&id=31882#toc (thanks Evan!)

Test Plan: No manual QA required

Reviewers: richsinn, EvanHahn

Reviewed By: richsinn, EvanHahn

Subscribers: jbbakst

Differential Revision: https://airtable.phacility.com/D10368