Support block names that start with a number in `block init`
Summary:
For https://airtable.com/tbl3f9cb8b2f15ab8/viw40mH4fGClHsR8K/rec3kE4W5yrDEKWf7?blocks=hide

Previously, creating a block that started with a number (eg "1Block") would error as we use that as the React component name, which must be a valid Javascript function.

Now, we append some text to the front to make it valid!

Added some tests that cover the above, as well as already-handled cases that "fix up" the text to make it a valid function name. (Also manually tested creating a block that started with a number).

Test Plan: No manual QA required

Reviewers: richsinn

Reviewed By: richsinn

Differential Revision: https://airtable.phacility.com/D10472