cli: switch long-polling timeout status code to non-error code to prevent dev console noise
Summary:
This switches the timeout response code for livereload polling requests from 408 Request Timeout (an error code) to 304 Not Modified (a success code). The error was causing a lot of noise in chrome's dev console, which this resolves:

{F124650}

Test Plan:
No QA needed.

Live-reload still works, but there are no console errors.

Reviewers: richsinn, jbbakst

Reviewed By: richsinn, jbbakst

Differential Revision: https://airtable.phacility.com/D10481