cli: block_server New copy msg for registration error
Summary:
Depends on D10497

Better copy message for registration errors. This will be shown on the block frame on a registration error. Registration errors are defined as if a block developer tries to run a local block from NOT the original base or block.

Tested manually.

Test Plan: No QA Needed

Reviewers: minicat

Reviewed By: minicat

Differential Revision: https://airtable.phacility.com/D10550