switch from lerna to yarn workspaces & release-it
Summary:
The saga continues...

This diff switches away from `lerna` to `yarn` workspaces and `release-it` for publishing.

Previously, we'd switched from `np` to `lerna` for publishing, for two reasons:
- we already needed to use `lerna` (or a tool like it) for installing dependencies
- `np` doesn't correctly tag releases for a monorepo. `np` would generate `v${version}` tags which weren't configurable. This would mean that we'd have the same tags for releases to difference packages, which wouldn't work.

For that second reason, we can't switch back to `np` now. `release-it` does all the same things `np` does, but is much more configurable - it lets us customise the tag format to support a monorepo

From a developer point of view the biggest changes are:
- no more `lerna` commands for managing dependencies - just use `yarn` as usual
- release using `cd packages/sdk` (or which ever package you want to publish) and `yarn release`. Follow the step-by-step guide.

I originally tried this when I was fire switching over to monorepos, and gave up because I couldn't get workspaces to play nice with flow. This is because yarn hoists all dependencies (apart from ones that might cause a conflict) up into the root node_modules. I resolved that by using the `nohoist` option to prevent hoisting of all our top-level dependencies. I did try this the first time around, but couldn't get it to work for some reason. In my defence, I was //incredibly// jetlagged at the time.

Test Plan: No QA.

Reviewers: EvanHahn, richsinn, jbbakst

Reviewed By: EvanHahn

Subscribers: #blocks

Differential Revision: https://airtable.phacility.com/D10548