sdk: replace `npm run` commands in package.json with `yarn run`
Summary: When I first set up the blocks-sdk repo, I used npm for running all these subcommands as it's sort-of the default and I wanted to make sure people with access to this repo could mess around with it even without yarn. Now we've switched to yarn workspaces, yarn is a hard requirement for working in this repo so we can use it here too.

Test Plan: -

Reviewers: EvanHahn, jbbakst, kasrak, richsinn

Reviewed By: kasrak, richsinn

Differential Revision: https://airtable.phacility.com/D10611