examples: move to root of directory using yarn workspaces
Summary:
This moves the example blocks out of `packages/sdk/example` and into their own top-level `examples` directory. The main motivation for this is simplifying the development experience for example blocks and tools like the ui playground added in https://airtable.phacility.com/D10918

It also adds them to yarn workspaces - that means that running `yarn` anywhere in the repo will install their dependencies.

They're pointed to the latest version of `@airtable/blocks`, which means yarn workspaces will symlink that dependency in rather than installing it from the registry - that means the examples now use the development version of the sdk.

To avoid issues with multiple versions of react, this moves the sdk's devDependency on react up a level to the root of the workspace. This should `import 'react';` in the sdk to resolve to the example's installed version of react.

Test Plan:
- `yarn` anywhere installs example dependencies
- `block run` in an example + `yarn watch` in the sdk causes block to live-reload when sdk code is changed

Reviewers: EvanHahn, jay-ransijn, maryrosecook

Reviewed By: EvanHahn, jay-ransijn

Differential Revision: https://airtable.phacility.com/D10929