examples: make sure examples are runnable out of the box
Summary:
There's a few changes here:
1. Make sure the remotes for every example point to a base in the shared 'Airtable blocks contstruction' workspace
2. Give the summary block a name and version in package.json so it gets picked up by yarn workspaces
3. Tweak yarns hoisting rules so that `react` and `react-dom` will get hoisted to the root of the workspace
4. Force examples to use the latest version of the sdk from npm

1 and 2 I think are fairly self explanatory. 3 and 4 are a bit weirder.

With the addition of 3, we can build blocks that consume the SDK directly from this repo - so long as they use the same version of react as the SDK, yarn workspaces will symlink everything together. The result is that `block run` will automatically live reload on changes to the sdk, allowing the development of blocks and the sdk in tandem in a way that's considerably less hacky and error-prone than the `--sdk-repo` option on blocks cli. This is enabled by the way yarn workspaces hoists dependencies and symlinks things together, it won't work for projects that are not part of the workspace in this repo. The main motivation for this change is to support D10918.

The problem is that the SDK uses an alpha version of react to test against - we need it to get `act(async () => {})` for automated tests. Ideally, our example blocks would get symlinked against the repo's version of the SDK so they're always the latest. But that means they have to use the same react version as the SDK - which means all our examples have to use an alpha version of react. We don't want that! Instead, in change 4, we just remove the examples from the yarn workspace for now. That means they won't get picked up by airbuild or have any symlinking done, but that's fine because we don't need any of that for now anyway.

Test Plan: All examples can run

Reviewers: jay-ransijn, maryrosecook, EvanHahn

Reviewed By: jay-ransijn

Differential Revision: https://airtable.phacility.com/D10964