Copy block_backend_wrapper to blocks-sdk/packages/cli.
Summary:
This diff does a blind copy of Hyperbase's `block_builder_service/deps/block_backend_wrapper` to `blocks-sdk/packages/cli`.

As part of the roadmap for releasing backend blocks, we will be making some changes to block_backend_wrapper, in particular
- Support for new block.json format (instead of reading from routes.json generated by block builder)
- Support for packaging user code in Lambda Layer

Since we still need to keep existing backend blocks working as-is, this diff creates a forked copy of `block_backend_wrapper` for use in `blocks-cli`. The fork is added to the blocks-sdk/packages/cli repo for now since it would initially only be used by the CLI and will be packaged along with the CLI when published to NPM.

Test Plan: No QA testing needed.

Reviewers: richsinn

Reviewed By: richsinn

Differential Revision: https://airtable.phacility.com/D10976