sdk: add storybook
Summary:
After trying to get Docz up and running in our setup, we finally forfeited and tried Storybook instead.
I have to give Storybook props because it gave us 0 issues in setting up.

Storybook is a UI dev environment we will be using for the presentational components that don't rely on being in a Block frame.

I created the Button stories as an example.

Test Plan:
- Run `yarn storybook`
- Verify it looks like the video

{F128429}

Reviewers: s2tephen

Reviewed By: s2tephen

Differential Revision: https://airtable.phacility.com/D11002