cli: Refactor blocks backend wrapper.
Summary:
This diff refactors the blocks backend wrapper, with the goal of supporting 1) usage in blocks CLI and 2) unit tests. To keep this diff to a pure refactor as much as possible, we refrain from making any application logic changes here.

The major changes are:

- The main application logic in `handle_event_async.js` is encapsulated into a class `BlocksBackendEventHandler` and moved to `blocks_backend_event_handler.js`.
- Environment-specific logic to import certain dependencies (`block.json`, backend SDK and user code) is extracted out to the AWS Lambda entrypoint `backend.js`, which would now pass the dependencies as constructor arguments to `BlocksBackendEventHandler`.
- Add flow types.
- Add some initial unit tests as a sanity check. Will add more detailed tests in subsequent diffs.

Test Plan: No QA testing needed.

Reviewers: richsinn

Reviewed By: richsinn

Differential Revision: https://airtable.phacility.com/D11101