sdk: add system configuration
Summary:
This diff adds a lot of configuration code. To see how it's actually used check out the `sdk-box` northstar branch.

The `system` folder contains configuration for properties that will be supported on the Box components, and others in the future.

## This diff is part of a series of stacking diffs:

1. `sdk-useStyledSystem`  https://airtable.phacility.com/D11141
2. `sdk-system` https://airtable.phacility.com/D11154 (current)
2. `sdk-boxComp` https://airtable.phacility.com/D11197

Test Plan: N/A. I will include steps once we have some more stuff that we can actually run.

Reviewers: s2tephen, somehats

Reviewed By: somehats

Subscribers: billylittlefield

Differential Revision: https://airtable.phacility.com/D11154