sdk: useStyledSystem
Summary:
This is the first diff in a series of diffs that will implement the Box UI primitive.

See  https://airtable.phacility.com/D11197  for how this will be used in practice.

## This diff is part of a series of stacking diffs:

1. `sdk-useStyledSystem`  https://airtable.phacility.com/D11141 (current)
2. `sdk-system` https://airtable.phacility.com/D11154
3. `sdk-boxComp` https://airtable.phacility.com/D11197

Test Plan:
N/A. I will include steps once we have some more stuff that we can actually run.

I can investigate how to add tests once we got some style parsers in a follow up diff, the stories that will be added in the future should also help with testing.

Reviewers: s2tephen, somehats

Reviewed By: s2tephen, somehats

Subscribers: billylittlefield

Differential Revision: https://airtable.phacility.com/D11141