sdk: upgrade react to latest stable version
Summary:
Previously, we were running an alpha build of react for our tests to get access to `async act` for testing async components. React 16.9.0 has that built in, so we can upgrade to that and remove our `actAsync` helper.

Depends on D11511

Test Plan: No QA needed. Tested with test block.

Reviewers: minicat

Reviewed By: minicat

Differential Revision: https://airtable.phacility.com/D11515