cli: Normalize path in chokidar 'ignored' for Windows
Summary:
In chokidar's [[https://github.com/paulmillr/chokidar#path-filtering | `ignored` attribute]] for path filtering, you can pass in a function to determine the path to ignore.
On Windows, for some reason, the argument that is passed into the function uses the forward slash `/` in the path, but the default segement separator for Windows is the backslash `\` (technically, Windows recognizes multiple path separators).

Therefore, fix is to use `path.normalize` on the argument that's passed in to make sure we're comparing normalized values.

Test Plan: no QA needed

Reviewers: jichu4n

Reviewed By: jichu4n

Differential Revision: https://airtable.phacility.com/D11637