cli: Explicitly invalidate Browserify cache and bundle on source code changes.
Summary:
Our current build loop relies on Watchify's `update` event to trigger bundling. This is very unfortunate because
  - We cannot tell ahead of time if a file will trigger bundling or not, because only changes files included in the frontend bundle will trigger bundling. As a result, we don't have a reliable signal for build completion, and currently rely on a hacky timeout-based method that is prone to race conditions.
  - It adds a significant amount of complexity to our build state management, since bundling is started asynchronously as opposed to directly after a transpile.
  - Watchify doesn't handle file deletions, because it only watches for the `change` event, not the `unlink` event: https://github.com/browserify/watchify/blob/bd2f677c448a87007915a17d5443255ba89a1123/index.js#L119
  - Watchify only fires `update` events after the first *successful* bundling. This requires us to additionally track whether the initial bundling has succeeded.

The reason why we currently rely on Watchify's `update` event is that it is an indication that Watchify has purged the changed file from its cache. If we trigger bundling before the `update` event fires, Browserify will end up using the outdated version of the transpiled file in the cache and produce an outdated (i.e. unchanged) bundle.

In this diff, we fix the root problem by purging changed files from Watchify's cache after transpilation, and explicitly triggering bundling. This allows us to significantly simplify the build loop, but to minimize changes we are deferring such refactors to a later diff.

Test Plan: No QA testing needed.

Reviewers: richsinn

Reviewed By: richsinn

Differential Revision: https://airtable.phacility.com/D11676