color_palette: handle zero width container
Summary: The color palette logs some errors when attempting to calculate square size in a container that has zero width, which is the case when the component initially mounts. This diff adds a validation in the square size calculation to just use the default size if "no squares will fit" (ie, container size is either zero or less than the size of the smallest possible square). No change in functionality, but clears up some errors.

Test Plan: qa can skip

Reviewers: jay-ransijn

Reviewed By: jay-ransijn

Differential Revision: https://airtable.phacility.com/D11792