sdk: hide overflow for blocks smaller than min size
Summary:
Followup to https://airtable.phacility.com/D11760 for the new SDK.

When showing the "Please make this block bigger" overlay in the blockframe, we want to ensure overflow content is hidden in the block.

To test this locally:
1. create a block whose scroll height/width exceeds the size of the block frame
2. add a min size on the viewport (`viewport.addMinSize({width: 400})`)
3. resize the block to be smaller than the viewport min size and trigger the overlay

Previously, the overlay would be scrollable and you could see the scrollable content of the block. Now, it is not scrollable.

Test Plan: QA can skip

Reviewers: s2tephen

Reviewed By: s2tephen

Differential Revision: https://airtable.phacility.com/D11785