record_card_list & collaborator_token: fix flow types
Summary:
Fixes two flow issues: once in CollaboratorToken and one in RecordCardList

Changes `CollaboratorTokenProps` in CollaboratorToken to use the `CollaboratorData` type instead of an adhoc object type.

Changes the `records` type in RecordCardList to use `Array<Record> | Array<RecordDef>` instead of `Array<Record | RecordDef>`, as this component shouldn't be used for mixing the two in one list.

A broader question is whether we actually need to support `Array<RecordDef>` at all in the UI.RecordCardList. I don't have the historical context to know whether this is a useful / planned use-case.

Test Plan: qa can skip

Reviewers: jay-ransijn, s2tephen

Reviewed By: jay-ransijn, s2tephen

Differential Revision: https://airtable.phacility.com/D11806