cli: Normalize backend route handler responses.
Summary:
This diff adds an additional step in the blocks backend wrapper that validates the responses returned from blocks backend route handlers.

The validation is based on the existing `normalize_user_response.js` module and reuses most of the same tests. The major differences are:

- Implemented using Joi instead of manual checks.

- Allows headers in object format.

When validation fails, a 500 error is returned along with a description of the error generated by Joi.

The normalization of the response body will be implemented in a subsequent diff.

Test Plan: No QA testing needed.

Reviewers: richsinn

Reviewed By: richsinn

Differential Revision: https://airtable.phacility.com/D11764